Use unique keys for Members.
You must use unique keys for items of a list if the list is shuffled or
otherwise heavily modified. In this case it's a MembersList component.
You can't use index of a list item 'cause it's modified on each render
call. Pass a timestamp based unique key inside the global state to sleep
well.
http://facebook.github.io/react/docs/multiple-components.html#dynamic-children