Add NavItemLink component.
Basically overwrite NavItem's onSelect event.
It's also known as onClick in other components but has this name
because NavItems never go alone and clicking it is selecting it
from a list.
The Navigation mix in provides this.transitionTo used to go
to the route.
The ActiveState mix in is used to check that current route
is the same as the NavItem's route and so the NavItem
is active. It doesn't work though.